+++ ext/makiri/rust/build.rs
@@ -237,7 +237,7 @@ fn main() {
         .allowlist_function("lxb_dom_node_destroy")
         .allowlist_function("lxb_dom_node_text_content_set")
         .allowlist_function("lxb_dom_element_set_attribute")
-        .allowlist_function("lxb_dom_element_remove_attribute")
+        .allowlist_function("lxb_dom_element_attr_by_name")
         .allowlist_function("lxb_dom_element_attr_append")
         .allowlist_function("lxb_dom_element_attr_remove")
         .allowlist_function("lxb_dom_attr_interface_create")